Struct ray_tracer_branch::math::vector::Vector3

pub struct Vector3 {
    pub x: f64,
    pub y: f64,
    pub z: f64,
}

A 3D vector with x, y, and z components.

Fields
x: f64

The x-coordinate of the vector.

y: f64

The y-coordinate of the vector.

z: f64

The z-coordinate of the vector.

Implementations

impl Vector3

pub fn new(x: f64, y: f64, z: f64) -> Self

Creates a new Vector3 instance with the specified x, y, and z components, and then normalizes the vector.

pub fn normalized(self) -> Self

Returns a normalized copy of the vector, i.e., a vector with the same direction but unit length.

pub fn cross(self, other: Self) -> Self

Returns the cross product of two vectors self and other.

pub fn dot(self, other: Vector3) -> f64

Returns the dot product of two vectors self and other.

pub fn mul(self, rhs: f64) -> Self

Multiplies each component of the vector self by the given scalar.

pub fn div(self, rhs: f64) -> Self

Divides each component of the vector self by the given scalar.

pub fn add(self, rhs: Self) -> Self

Adds the vector other to self.

pub fn sub(self, rhs: Self) -> Self

Subtracts the vector other from self.

pub fn neg(self) -> Self

Returns the negation of the vector self.

pub fn length(&self) -> f64

Returns the length of the vector.

pub fn length_squared(&self) -> f64

Returns the squared length of the vector.

pub fn theta(a: Vector3, b: Vector3) -> f64

Returns the angle (in radians) between two normalized vectors a and b.

pub fn distance_to(&self, other: Vector3) -> f64

Returns the distance between the vector self and another vector other.

pub fn random(min: f64, max: f64) -> Vector3

Generates a random 3D vector within the specified range [min, max].

pub fn random_unit() -> Vector3

Generates a random 3D unit vector (vector with unit length).

pub fn random_in_unit_sphere() -> Vector3

Generates a random 3D vector within a unit sphere.

pub fn reflect(i: Vector3, n: Vector3) -> Vector3

Reflects the incident vector i off a surface with the given normal vector n.